refactor, factory method pattern
