+++ src/Card/Generator/CardUsaFactory.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * Licensed under The GPL-3.0 License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @author   Christopher Castro <chris@quickapps.es>
+ * @license  http://opensource.org/licenses/gpl-3.0.html GPL-3.0 License
+ */
+namespace Bingo\Card\Generator;
+
+use Bingo\Card\CardInterface;
+
+class CardUsaFactory implements CardFactoryInterface
+{
+
+    /**
+     * {@inheritdoc}
+     */
+    public function build(): CardInterface
+    {
+    }
+}
